backend/requirements.txt
fastapi
uvicorn
openai
pydantic
pydantic-settings
python-dotenv
pytest
python-jose[cryptography]
passlib[bcrypt]